move all implementations in Fan.cpp to Fan.h, and remove Fan.cpp entirely
